refactor: move getTextWidth to utils/index
